# [WIP] wmma: loosen tensor core triggering criteria