--- a/test/test_linearizer.py
+++ b/test/test_linearizer.py
@@ -112,7 +112,7 @@ class TestLinearizer(unittest.TestCase):
       r = a.matmul(b, acc_dtype=tc.dtype_out)
       realized_ast, _ = helper_realized_ast(r)
       k = Linearizer(realized_ast)
-      k.apply_tensor_cores(1)
+      k.apply_tensor_cores(use_tensor_cores=1, axis=0)
       k.linearize()
       assert len([uop for uop in k.uops if uop.uop == UOps.WMMA]) == 1, "tensor core not triggered"
       np_c = np_a @ np_b
@@ -394,7 +394,7 @@ def helper_linearizer_opt(r:Tensor, opts=[], apply_tc=False, atol=1e-4, rtol=1e-
   def check_opt(opts, create_k, to_prg):
     k = create_k()
     if apply_tc:
-      assert k.apply_tensor_cores(1, opts), "no tensor core triggered"
+      assert k.apply_tensor_cores(use_tensor_cores=1, axis=0, extra_opts=opts), "no tensor core triggered"
     else:
       for opt in opts:
         k.apply_opt(opt)
@@ -543,7 +543,7 @@ class TestLinearizerOpts(unittest.TestCase):
     for x in invalid_opts:
       k = Linearizer(realized_ast)
       with self.assertRaises(AssertionError):
-        assert k.apply_tensor_cores(use_tensor_cores=1, extra_opts=x), "no valid tensor core" # for METAL in runners
+        assert k.apply_tensor_cores(use_tensor_cores=1, axis=0, extra_opts=x), "no valid tensor core" # for METAL in runners
 
   def test_tensor_core_opts(self):
     if not Device[Device.DEFAULT].compiler.linearizer_opts.has_local: