when using BEAM search, loosen the tensor core criteria to allow the search to attempt to apply it to more kernels.

for hlb_cifar, this allows BEAM to go from using 6 kernels that have tensor cores applied to up to 18 kernels while using BEAM, increasing the performance with (32ms to 19ms steps) and without WINO (37ms to 35ms steps).

Depends on #3275